ASALIGHT-160/161: Update SDK and SampleApp to use 15.04 Core lifecycle APIs ASALIGHT-165/166: Build iOS SDK and SampleApp using XCode 6 and 64-bit native AJSI-605/608: Continued implementation of Lighting SDK 1.1

+++ helper-lib/objc/LSFObjCHelper/LSFObjCHelper/LSFGroupModel.h
@@ -26,7 +26,7 @@
 @property (nonatomic, strong) NSSet *lamps;
 @property (nonatomic, strong) NSSet *groups;
 @property (nonatomic) int duplicates;
-@property (nonatomic) unsigned int delay;
+@property (nonatomic) unsigned long delay;
 @property (nonatomic) unsigned int groupColorTempMin;
 @property (nonatomic) unsigned int groupColorTempMax;
